% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggSketch.R
\name{ggSketch}
\alias{ggSketch}
\title{Generate a sketch-style plot from a photo}
\usage{
ggSketch(photopath, color = "#000000", linewidth = 0.3)
}
\arguments{
\item{photopath}{Character
The path of photo}

\item{color}{Character
The color of sketch plot}

\item{linewidth}{Numeric
The linewidth of sketch plot}
}
\value{
a ggplot object
}
\description{
Generate a sketch-style plot from a photo
}
\examples{
NULL
}
