% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_photo.R
\name{parse_photo}
\alias{parse_photo}
\title{Parse a photo and extract contour coordinates}
\usage{
parse_photo(photopath)
}
\arguments{
\item{photopath}{Character
The path of photo}
}
\value{
A data frame
}
\description{
Parse a photo and extract contour coordinates
}
\examples{
NULL
}
