% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-arrowhead.R
\name{sketch_arrowheads}
\alias{sketch_arrowheads}
\title{The available arrowhead styles}
\usage{
sketch_arrowheads()
}
\value{
A character vector of arrowhead names.
}
\description{
Valid values for the \code{head} argument of \code{\link[=geom_sketch_arrow]{geom_sketch_arrow()}},
\code{\link[=annotate_sketch_arrow]{annotate_sketch_arrow()}}, \code{\link[=geom_sketch_callout]{geom_sketch_callout()}} and the \code{style} of
\code{\link[=arrowhead]{arrowhead()}}.
}
\examples{
sketch_arrowheads()
}
\seealso{
Other sketch-core:
\code{\link[=arrowhead]{arrowhead()}},
\code{\link[=curve_fill]{curve_fill()}},
\code{\link[=engrave_fill]{engrave_fill()}},
\code{\link[=engrave_ladder]{engrave_ladder()}},
\code{\link[=hachure_fill]{hachure_fill()}},
\code{\link[=hachure_fill_multi]{hachure_fill_multi()}},
\code{\link[=repel_layout]{repel_layout()}},
\code{\link[=rough_arc]{rough_arc()}},
\code{\link[=rough_bezier]{rough_bezier()}},
\code{\link[=rough_ellipse]{rough_ellipse()}},
\code{\link[=roughen_polyline]{roughen_polyline()}},
\code{\link[=sketch_fill]{sketch_fill()}},
\code{\link[=sketch_fill_multi]{sketch_fill_multi()}},
\code{\link[=spray_scatter]{spray_scatter()}},
\code{\link[=stroke_profile]{stroke_profile()}},
\code{\link[=stroke_ribbon]{stroke_ribbon()}},
\code{\link[=treemap_layout]{treemap_layout()}},
\code{\link[=wash_bleed]{wash_bleed()}},
\code{\link[=watercolor_wash]{watercolor_wash()}},
\code{\link[=watercolor_wash_multi]{watercolor_wash_multi()}}
}
\concept{sketch-core}
