% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketch-style.R
\name{sketch_styles}
\alias{sketch_styles}
\title{The available style presets}
\usage{
sketch_styles()
}
\value{
A character vector of style names.
}
\description{
The valid values for \code{\link[=sketch_style]{sketch_style()}}.
}
\examples{
sketch_styles()
}
\seealso{
Other sketch-style:
\code{\link[=sketch_style]{sketch_style()}}
}
\concept{sketch-style}
