% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-sketch-reference.R
\name{GeomSketchAbline}
\alias{GeomSketchAbline}
\alias{GeomSketchHline}
\alias{GeomSketchVline}
\alias{geom_sketch_abline}
\alias{geom_sketch_hline}
\alias{geom_sketch_vline}
\title{Sketchy reference lines}
\usage{
GeomSketchAbline

GeomSketchHline

GeomSketchVline

geom_sketch_abline(
  mapping = NULL,
  data = NULL,
  ...,
  slope,
  intercept,
  roughness = 0.6,
  bowing = 0.5,
  n_passes = 2L,
  seed = NULL,
  na.rm = FALSE,
  show.legend = NA
)

geom_sketch_hline(
  mapping = NULL,
  data = NULL,
  ...,
  yintercept,
  roughness = 0.6,
  bowing = 0.5,
  n_passes = 2L,
  seed = NULL,
  na.rm = FALSE,
  show.legend = NA
)

geom_sketch_vline(
  mapping = NULL,
  data = NULL,
  ...,
  xintercept,
  roughness = 0.6,
  bowing = 0.5,
  n_passes = 2L,
  seed = NULL,
  na.rm = FALSE,
  show.legend = NA
)
}
\arguments{
\item{mapping, data, ...}{Standard layer arguments. Usually omitted in favour of
the intercept arguments below.}

\item{slope, intercept}{For \code{geom_sketch_abline()}.}

\item{roughness, bowing, n_passes, seed}{Sketch parameters. Reference lines
default to a gentle \code{roughness = 0.6}.}

\item{na.rm, show.legend}{Standard layer arguments.}

\item{yintercept}{For \code{geom_sketch_hline()}.}

\item{xintercept}{For \code{geom_sketch_vline()}.}
}
\value{
A \code{ggplot2} layer object.
}
\description{
Hand-drawn \code{abline} / \code{hline} / \code{vline} reference lines that span the panel -
the sketch analogues of \code{\link[ggplot2:geom_abline]{ggplot2::geom_abline()}}, \code{\link[ggplot2:geom_hline]{ggplot2::geom_hline()}}, and
\code{\link[ggplot2:geom_vline]{ggplot2::geom_vline()}}. As with ggplot2, you usually pass the intercepts as
arguments rather than mapping them.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) +
  geom_sketch_point(seed = 1L) +
  geom_sketch_hline(yintercept = 20, colour = "red", seed = 2L) +
  geom_sketch_vline(xintercept = 3, colour = "blue", seed = 3L) +
  theme_sketch()
}
\seealso{
Other sketch-geoms:
\code{\link{GeomSketchArrow}},
\code{\link{GeomSketchBoxplot}},
\code{\link{GeomSketchBracket}},
\code{\link{GeomSketchCallout}},
\code{\link{GeomSketchCol}},
\code{\link{GeomSketchContourFilled}},
\code{\link{GeomSketchCurve}},
\code{\link{GeomSketchDotplot}},
\code{\link{GeomSketchDumbbell}},
\code{\link{GeomSketchEdge}},
\code{\link{GeomSketchEllipse}},
\code{\link{GeomSketchEngrave}},
\code{\link{GeomSketchHex}},
\code{\link{GeomSketchLine}},
\code{\link{GeomSketchLinerange}},
\code{\link{GeomSketchLollipop}},
\code{\link{GeomSketchMarkCircle}},
\code{\link{GeomSketchMarkHull}},
\code{\link{GeomSketchPath}},
\code{\link{GeomSketchPoint}},
\code{\link{GeomSketchPolygon}},
\code{\link{GeomSketchRect}},
\code{\link{GeomSketchRibbon}},
\code{\link{GeomSketchRug}},
\code{\link{GeomSketchSegment}},
\code{\link{GeomSketchSfPolygon}},
\code{\link{GeomSketchSlope}},
\code{\link{GeomSketchSmooth}},
\code{\link{GeomSketchSpoke}},
\code{\link{GeomSketchViolin}},
\code{\link{StatSketchBeeswarm}},
\code{\link{StatSketchCalendar}},
\code{\link{StatSketchDensityRidges}},
\code{\link{StatSketchPie}},
\code{\link{StatSketchRadar}},
\code{\link{StatSketchStream}},
\code{\link{StatSketchTreemap}},
\code{\link{StatSketchWaffle}},
\code{\link[=annotate_sketch]{annotate_sketch()}},
\code{\link[=annotate_sketch_arrow]{annotate_sketch_arrow()}},
\code{\link[=annotate_sketch_callout]{annotate_sketch_callout()}},
\code{\link[=geom_sketch_alluvial]{geom_sketch_alluvial()}},
\code{\link[=geom_sketch_bin2d]{geom_sketch_bin2d()}},
\code{\link[=geom_sketch_chord]{geom_sketch_chord()}},
\code{\link[=geom_sketch_contour]{geom_sketch_contour()}},
\code{\link[=geom_sketch_count]{geom_sketch_count()}},
\code{\link[=geom_sketch_density]{geom_sketch_density()}},
\code{\link[=geom_sketch_density2d]{geom_sketch_density2d()}},
\code{\link[=geom_sketch_ecdf]{geom_sketch_ecdf()}},
\code{\link[=geom_sketch_function]{geom_sketch_function()}},
\code{\link[=geom_sketch_histogram]{geom_sketch_histogram()}},
\code{\link[=geom_sketch_jitter]{geom_sketch_jitter()}},
\code{\link[=geom_sketch_parallel]{geom_sketch_parallel()}},
\code{\link[=geom_sketch_qq]{geom_sketch_qq()}},
\code{\link[=geom_sketch_quantile]{geom_sketch_quantile()}},
\code{\link[=geom_sketch_text]{geom_sketch_text()}},
\code{\link[=sketch_graph]{sketch_graph()}}
}
\concept{sketch-geoms}
