% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-sketch-interval.R
\name{GeomSketchLinerange}
\alias{GeomSketchLinerange}
\alias{GeomSketchPointrange}
\alias{GeomSketchErrorbar}
\alias{GeomSketchCrossbar}
\alias{geom_sketch_linerange}
\alias{geom_sketch_pointrange}
\alias{geom_sketch_errorbar}
\alias{geom_sketch_crossbar}
\title{Sketchy interval geoms}
\usage{
GeomSketchLinerange

GeomSketchPointrange

GeomSketchErrorbar

GeomSketchCrossbar

geom_sketch_linerange(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  roughness = 0.7,
  bowing = 1,
  n_passes = 2L,
  seed = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_sketch_pointrange(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  fatten = 4,
  roughness = 0.7,
  bowing = 1,
  n_passes = 2L,
  seed = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_sketch_errorbar(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  width = 0.5,
  roughness = 0.7,
  bowing = 1,
  n_passes = 2L,
  seed = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_sketch_crossbar(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  width = 0.5,
  roughness = 0.7,
  bowing = 1,
  n_passes = 2L,
  seed = NULL,
  fill_style = "solid",
  hachure_angle = 45,
  hachure_gap = NULL,
  fill_weight = 0.5,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{Standard
layer arguments.}

\item{roughness, bowing, n_passes, seed}{Sketch parameters.}

\item{fatten}{For \code{pointrange}, multiply the point \code{size} by this factor.}

\item{width}{For \code{errorbar}/\code{crossbar}, the cap/box width in data units.}

\item{fill_style, hachure_angle, hachure_gap, fill_weight}{Fill parameters for
the \code{crossbar} box.}
}
\value{
A \code{ggplot2} layer object.
}
\description{
Hand-drawn uncertainty intervals at each \code{x}: \code{geom_sketch_linerange()} draws
a vertical range from \code{ymin} to \code{ymax}; \code{geom_sketch_pointrange()} adds a point
at \code{y}; \code{geom_sketch_errorbar()} adds end caps; \code{geom_sketch_crossbar()} draws
a rough box with a line at \code{y}. Sketch analogues of
\code{\link[ggplot2:geom_linerange]{ggplot2::geom_linerange()}} and friends. Vertical orientation only in this
release.
}
\examples{
library(ggplot2)
df <- data.frame(x = c("a", "b", "c"), y = c(2, 5, 4),
                 lo = c(1, 4, 2.5), hi = c(3, 6, 5.5))
ggplot(df, aes(x, y)) +
  geom_sketch_pointrange(aes(ymin = lo, ymax = hi), seed = 1L) +
  theme_sketch()
}
\seealso{
Other sketch-geoms:
\code{\link{GeomSketchAbline}},
\code{\link{GeomSketchArrow}},
\code{\link{GeomSketchBoxplot}},
\code{\link{GeomSketchBracket}},
\code{\link{GeomSketchCallout}},
\code{\link{GeomSketchCol}},
\code{\link{GeomSketchContourFilled}},
\code{\link{GeomSketchCurve}},
\code{\link{GeomSketchDotplot}},
\code{\link{GeomSketchDumbbell}},
\code{\link{GeomSketchEdge}},
\code{\link{GeomSketchEllipse}},
\code{\link{GeomSketchEngrave}},
\code{\link{GeomSketchHex}},
\code{\link{GeomSketchLine}},
\code{\link{GeomSketchLollipop}},
\code{\link{GeomSketchMarkCircle}},
\code{\link{GeomSketchMarkHull}},
\code{\link{GeomSketchPath}},
\code{\link{GeomSketchPoint}},
\code{\link{GeomSketchPolygon}},
\code{\link{GeomSketchRect}},
\code{\link{GeomSketchRibbon}},
\code{\link{GeomSketchRug}},
\code{\link{GeomSketchSegment}},
\code{\link{GeomSketchSfPolygon}},
\code{\link{GeomSketchSlope}},
\code{\link{GeomSketchSmooth}},
\code{\link{GeomSketchSpoke}},
\code{\link{GeomSketchViolin}},
\code{\link{StatSketchBeeswarm}},
\code{\link{StatSketchCalendar}},
\code{\link{StatSketchDensityRidges}},
\code{\link{StatSketchPie}},
\code{\link{StatSketchRadar}},
\code{\link{StatSketchStream}},
\code{\link{StatSketchTreemap}},
\code{\link{StatSketchWaffle}},
\code{\link[=annotate_sketch]{annotate_sketch()}},
\code{\link[=annotate_sketch_arrow]{annotate_sketch_arrow()}},
\code{\link[=annotate_sketch_callout]{annotate_sketch_callout()}},
\code{\link[=geom_sketch_alluvial]{geom_sketch_alluvial()}},
\code{\link[=geom_sketch_arc_diagram]{geom_sketch_arc_diagram()}},
\code{\link[=geom_sketch_bin2d]{geom_sketch_bin2d()}},
\code{\link[=geom_sketch_bump]{geom_sketch_bump()}},
\code{\link[=geom_sketch_chord]{geom_sketch_chord()}},
\code{\link[=geom_sketch_contour]{geom_sketch_contour()}},
\code{\link[=geom_sketch_count]{geom_sketch_count()}},
\code{\link[=geom_sketch_dendrogram]{geom_sketch_dendrogram()}},
\code{\link[=geom_sketch_density]{geom_sketch_density()}},
\code{\link[=geom_sketch_density2d]{geom_sketch_density2d()}},
\code{\link[=geom_sketch_ecdf]{geom_sketch_ecdf()}},
\code{\link[=geom_sketch_function]{geom_sketch_function()}},
\code{\link[=geom_sketch_histogram]{geom_sketch_histogram()}},
\code{\link[=geom_sketch_jitter]{geom_sketch_jitter()}},
\code{\link[=geom_sketch_marimekko]{geom_sketch_marimekko()}},
\code{\link[=geom_sketch_mosaic]{geom_sketch_mosaic()}},
\code{\link[=geom_sketch_parallel]{geom_sketch_parallel()}},
\code{\link[=geom_sketch_qq]{geom_sketch_qq()}},
\code{\link[=geom_sketch_quantile]{geom_sketch_quantile()}},
\code{\link[=geom_sketch_rose]{geom_sketch_rose()}},
\code{\link[=geom_sketch_sunburst]{geom_sketch_sunburst()}},
\code{\link[=geom_sketch_text]{geom_sketch_text()}},
\code{\link[=sketch_graph]{sketch_graph()}}
}
\concept{sketch-geoms}
